refactor(types): make types package runtime-free and recentralise ids
The types package CONVENTIONS state it must be runtime-free, yet two
runtime values had leaked in:

- DEFAULT_SESSION_PROVIDER_PREFERENCE: moved to packages/core providers/
  preferences.ts. store.ts now imports it from @kay-am/core. The db
  migration runner test inlines a local copy to avoid the forbidden
  db → core upward dep.
- CONFIG_BUNDLE_SCHEMA_VERSION: was only consumed via typeof inside the
  types package itself; replaced with a literal ConfigBundleSchemaVersion
  type. The Rust side already owns the canonical value.

Brand types were also scattered:
- PlanId and PlanConsumptionId moved from plan.ts to ids.ts; plan.ts
  now imports them like every other domain type does.

Domain types previously defined inside @kay-am/db queries are now first
class in @kay-am/types:
- notification.ts holds Notification, NotificationKind, NotificationSeverity
- workspace-init-script.ts holds WorkspaceInitScript

The db query files re-export the types under their old names so existing
deep imports keep compiling; new code should import from @kay-am/types.

Net: types package is back to types-only, ids are centralised, domain
types are accessible without reaching into a sibling package's queries.
