+++ packages/types/src/provider-preference.ts
@@ -10,8 +10,3 @@ export interface TurnProviderOverride {
   readonly providerId: ProviderId;
   readonly model?: string;
 }
-
-export const DEFAULT_SESSION_PROVIDER_PREFERENCE: SessionProviderPreference = {
-  defaultProvider: 'anthropic',
-  allowTurnOverride: true,
-};
