+++ packages/types/src/workspace-init-script.ts
@@ -1,0 +1,8 @@
+import type { IsoDateTime, WorkspaceId } from './ids';
+
+export interface WorkspaceInitScript {
+  readonly id: string;
+  readonly workspaceId: WorkspaceId;
+  readonly content: string;
+  readonly createdAt: IsoDateTime;
+}
